implementando carrito de compras

+++ lib/core/utils/math.dart
@@ -1,5 +1,5 @@
 String calculatePercentageChange(double lastPrice, double currentPrice) {
-  int percentageChange = ((lastPrice - currentPrice) / lastPrice).toInt() * 100;
+  double percentageChange = ((lastPrice - currentPrice) / lastPrice) * 100;
 
   if (percentageChange > 0) {
     return '${percentageChange.round()}%';
